pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "MCP Development Learning Project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=0.1.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "bs4>=0.0.2",
    "requests>=2.32.5",
    "python-dotenv>=1.2.1",
]
